+++ lib/ffi/inliner/compilers/tcc.rb
@@ -18,7 +18,7 @@ class TCC < Compiler
 
   def input
     File.join(Inliner.directory, "#{digest}.c").tap {|path|
-      File.open(path, 'w') { |f| f.write(@code) }
+      File.open(path, 'w') { |f| f.write(@code) } unless File.exists?(path)
     }
   end
 
@@ -31,6 +31,8 @@ class TCC < Compiler
   end
 
   def compile
+    return output if File.exists?(output)
+
     unless system(if RbConfig::CONFIG['target_os'] =~ /mswin|mingw/
       "tcc -rdynamic -shared #{libs} -o \"#{output}\" \"#{input}\" 2>\"#{log}\""
     else
